refactor: refactor data fetch impl.
